ffi: fix poll() return value in include file
Fixes #748.

+++ include/quiche.h
@@ -511,8 +511,8 @@ enum quiche_h3_event_type {
 typedef struct Http3Event quiche_h3_event;
 
 // Processes HTTP/3 data received from the peer.
-int quiche_h3_conn_poll(quiche_h3_conn *conn, quiche_conn *quic_conn,
-                        quiche_h3_event **ev);
+int64_t quiche_h3_conn_poll(quiche_h3_conn *conn, quiche_conn *quic_conn,
+                            quiche_h3_event **ev);
 
 // Returns the type of the event.
 enum quiche_h3_event_type quiche_h3_event_type(quiche_h3_event *ev);
